refactor: shared base for load and deploy
